Fix bug: Item Customizer does not allow Weight of less than 1lb Issue#: 1947360
git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@6067 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/pcgen/core/UnitSet.java
@@ -303,13 +303,13 @@ public final class UnitSet
 	 * Convert weight in units to weight in pounds.
 	 * 
 	 * @param weight
-	 * @return int weight in pounds.
+	 * @return double weight in pounds.
 	 */
-	public int convertWeightFromUnitSet(final double weight)
+	public double convertWeightFromUnitSet(final double weight)
 	{
 		final double weightInPounds = weight / getWeightFactor();
 
-		return (int) weightInPounds;
+		return weightInPounds;
 	}
 
 	/**
